split out a build command from the synapsectl deploy command
