+++ synapse/cli/__main__.py
@@ -1,13 +1,23 @@
 #!/usr/bin/env python
 import argparse
-import logging
 import ipaddress
+import logging
 import sys
-
 from importlib import metadata
-from synapse.cli import discover, rpc, streaming, offline_plot, files, deploy, taps
-from rich.logging import RichHandler
+
 from rich.console import Console
+from rich.logging import RichHandler
+
+from synapse.cli import (
+    build,
+    deploy,
+    discover,
+    files,
+    offline_plot,
+    rpc,
+    streaming,
+    taps,
+)
 from synapse.utils.discover import find_device_by_name
 
 
@@ -66,6 +76,7 @@ def main():
     files.add_commands(subparsers)
     taps.add_commands(subparsers)
     deploy.add_commands(subparsers)
+    build.add_commands(subparsers)
     args = parser.parse_args()
 
     # If we need to setup the device URI, do that now
